backend/requirements.txt
absl-py
alembic
annotated-types
anyio
asgi-lifespan
astunparse
async-timeout
asyncpg
bcrypt
beautifulsoup4
boto3
botocore
certifi
charset-normalizer
click
contourpy
cycler
dnspython
ecdsa
email_validator
exceptiongroup
fastapi
filelock
finbert-embedding
flatbuffers
fonttools
frozendict
fsspec
gast
google-pasta
greenlet
grpcio
h11
h5py
html5lib
httpcore
httpx
httptools
huggingface-hub
idna
importlib_resources
iniconfig
Jinja2
jmespath
joblib
keras
kiwisolver
libclang
lxml
Mako
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml-dtypes
mpmath
multitasking
namex
networkx
nltk
numpy
opt-einsum
optree
packaging
pandas
passlib[bcrypt]
peewee
pillow
platformdirs
pluggy
protobuf
psycopg2-binary
pyasn1
pydantic
pydantic-settings
pydantic_core
Pygments
pyparsing
pytest
pytest-asyncio
python-dateutil
python-dotenv
python-jose
python-multipart
pytorch-pretrained-bert
pytz
PyYAML
redis
regex
requests
respx
rich
rsa
s3transfer
safetensors
scipy
six
sniffio
soupsieve
SQLAlchemy
starlette
sympy
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
textblob
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typing_extensions
tzdata
urllib3
uvicorn
uvloop
watchfiles
webencodings
websockets
Werkzeug
wrapt
yfinance
zipp
tf-keras